implemented VertxBuilder.VertxBuilderBridge as implementation to io.vertx.core.VertxBuilder and possibility to create io.vertx.core.impl.VertxBuilder from it
related code cleanup and removed code duplication
